Stop using `Clock` to find current term
To resolve #6, we need to find the current term without caring what
today's date it. It was a fragile and bad implementation anyway.

This commit sorts all of the terms for a given Member of Congress. The
`mostRecent` term is considered to be the current term for active
legislators.

I performed a text diff between `trunk` and this commit for program
output. They were identical.

References #6